refactor: extract PackageInfo to Project mapping
